agentic/requirements.txt
openai>=1.12.0
pdfminer.six>=20221105
pydantic>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
